move pixel resize family to separated source file
